requirements.txt
base58==2.1.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
grpcio==1.64.0
grpcio-tools==1.64.0
idna==3.7
ipfshttpclient==0.7.0
iroha==1.6.0.2
multiaddr==0.0.9
netaddr==1.3.0
pip==20.0.2
pkg-resources==0.0.0
protobuf==3.20.1
pycparser==2.22
pycryptodome==3.20.0
PyNaCl==1.5.0
requests==2.32.3
setuptools==44.0.0
six==1.16.0
urllib3==2.2.1
varint==1.0.2